feature-gate `netconf` transports and unify dependencies
