+++ Cargo.toml
@@ -16,22 +16,31 @@ categories = ["network-programming", "database"]
 [workspace.dependencies]
 # internal
 bgpfu-lib = { path = "./lib", version = "0.1.0-alpha.2" }
-bgpfu-netconf = { path = "./netconf", version = "0.1.0-alpha.2" }
+bgpfu-netconf = { path = "./netconf", version = "0.1.0-alpha.2", default-features = false }
 # external
 anyhow = "^1.0"
+async-trait = "^0.1"
+bytes = "^1.0"
 clap = {version = "^4.0", features = ["derive"]}
 clap-verbosity-flag = "^2.0"
 chrono = "^0.4"
 generic-ip = "^0.1"
+iri-string = "^0.7"
 irrc = "^0.1"
+memchr = "^2.0"
+paste = "^1.0"
 quick-xml = "^0.31"
 rpsl = "^0.1"
+russh = "^0.39"
+russh-keys = "^0.38"
 rustls-pemfile = "^2.0"
 rustls-pki-types = "^1.0"
 thiserror = "^1.0"
-tokio = { version = "^1.0", features = ["rt-multi-thread", "fs", "macros", "signal"] }
+tokio = { version = "^1.0", default-features = false }
+tokio-rustls = "^0.25"
 tracing = { version = "^0.1", features = ["log"] }
 tracing-appender = "^0.2"
 tracing-log = "^0.2"
 tracing-subscriber = "^0.3"
+uuid = { version = "^1.0", features = ["v4", "fast-rng"] }
 version-sync = "^0.9"
