+++ junos-agent/Cargo.toml
@@ -16,7 +16,6 @@ name = "bgpfu-junos-agent"
 [dependencies]
 anyhow.workspace = true
 bgpfu-lib.workspace = true
-bgpfu-netconf = { workspace = true, features = ["junos"] }
 clap.workspace = true
 clap-verbosity-flag.workspace = true
 generic-ip.workspace = true
@@ -24,11 +23,12 @@ quick-xml.workspace = true
 rpsl.workspace = true
 rustls-pemfile.workspace = true
 rustls-pki-types.workspace = true
-tokio.workspace = true
 tracing.workspace = true
 tracing-appender.workspace = true
 tracing-log.workspace = true
 tracing-subscriber.workspace = true
+bgpfu-netconf = { workspace = true, features = ["tls", "junos"] }
+tokio = { workspace = true, features = ["fs", "signal", "time", "rt-multi-thread", "macros"] }
 
 [dev-dependencies]
 version-sync.workspace = true
