+++ netconf/src/transport/mod.rs
@@ -1,18 +1,16 @@
-use std::{
-    convert::Infallible,
-    fmt::{self, Debug},
-    str::FromStr,
-};
-
 use async_trait::async_trait;
 use bytes::Bytes;
 
 use crate::Error;
 
+#[cfg(feature = "ssh")]
 mod ssh;
-pub use self::ssh::Ssh;
+#[cfg(feature = "ssh")]
+pub use self::ssh::{Password, Ssh};
 
+#[cfg(feature = "tls")]
 mod tls;
+#[cfg(feature = "tls")]
 pub use self::tls::Tls;
 
 pub trait Transport: Send {
@@ -30,27 +28,4 @@ pub trait SendHandle: Send {
 #[async_trait]
 pub trait RecvHandle: Send {
     async fn recv(&mut self) -> Result<Bytes, Error>;
-}
-
-#[derive(Clone)]
-pub struct Password(String);
-
-impl Password {
-    #[must_use]
-    pub fn into_inner(self) -> String {
-        self.0
-    }
-}
-
-impl Debug for Password {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_tuple("Password").field(&"****").finish()
-    }
-}
-
-impl FromStr for Password {
-    type Err = Infallible;
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Ok(Self(s.to_string()))
-    }
 }
